refactor(server): extract core route registry
